Split iOS review managed media
